requirements.txt
# ─────────────────────────────────────────────────────────────────
#  LLM-Powered SOC Analyst — Python Requirements
# ─────────────────────────────────────────────────────────────────

# ── FastAPI Backend ───────────────────────────────────────────────
fastapi==0.115.12
uvicorn==0.34.0

# ── Google Gemini LLM ─────────────────────────────────────────────
# New google-genai SDK (used in llm_agent.py)
google-genai==1.9.0

# ── RAG Framework (LangChain) ─────────────────────────────────────
langchain==0.3.21
langchain-community==0.3.20
langchain-core==0.3.48
langchain-huggingface==0.1.2

# ── Vector Database ───────────────────────────────────────────────
chromadb==0.6.3

# ── Sentence Embeddings ───────────────────────────────────────────
sentence-transformers==3.4.1

# ── Environment Variables ─────────────────────────────────────────
python-dotenv==1.1.0

# ── Machine Learning — LSTM ───────────────────────────────────────
torch==2.6.0
torchvision==0.21.0

# ── Graph Analysis ────────────────────────────────────────────────
networkx==3.4.2

# ── Data Processing ───────────────────────────────────────────────
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.6.1

# ── Validation and Models ─────────────────────────────────────────
pydantic==2.11.1

# ── HTTP Requests ─────────────────────────────────────────────────
requests==2.32.3

# ── JSON/utility ─────────────────────────────────────────────────
orjson==3.10.16
rich==13.9.4

# ── Visualization (for training curves — optional) ───────────────
matplotlib==3.10.0
